# internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ccw",
    "version": "1",
    "description": "One base URL for many LLM providers. The model in the request body selects the provider: send \"<provider>/<model>\", or a bare model id that one provider serves. OpenAI Chat Completions, Anthropic Messages and OpenAI Responses callers use the same URL. Each caller gets answers and errors in its own API shape."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "bearer": [] }, { "apiKey": [] }],
  "paths": {
    "/v1/models": {
      "get": {
        "summary": "List the models that an active account serves",
        "responses": {
          "200": { "description": "The model list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelList" } } } },
          "401": { "$ref": "#/components/responses/OpenAIError" }
        }
      }
    },
    "/v1/models/{model}": {
      "get": {
        "summary": "Read one model",
        "parameters": [{
          "name": "model", "in": "path", "required": true,
          "description": "\"<provider>/<model>\", or a bare id that one provider serves. The slash can also be sent as %2F.",
          "schema": { "type": "string" }
        }],
        "responses": {
          "200": { "description": "The model", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } } },
          "404": { "$ref": "#/components/responses/OpenAIError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "description": "The master token, or a key made in the dashboard (sk-ccw-...)" },
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key", "description": "The same token, in the header that Anthropic clients send" }
    },
    "headers": {
      "RequestId": { "description": "One id per call. When the provider sends its own request id, that id replaces this one.", "schema": { "type": "string" } }
    },
    "responses": {
      "OpenAIError": {
        "description": "An error in the OpenAI envelope",
        "headers": { "Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpenAIError" } } }
      },
...
    },
    "schemas": {
      "OpenAIError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["message", "type", "param", "code"],
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string" },
              "param": { "type": ["string", "null"] },
              "code": { "type": ["string", "null"], "examples": ["invalid_api_key", "model_not_found", "rate_limit_exceeded"] }
            }
          }
        }
      },
...
      "Model": {
        "type": "object",
        "description": "Both vocabularies in one object: OpenAI reads object, created and owned_by. Anthropic reads type, display_name and created_at. created is the time ccw read the provider list.",
        "required": ["id", "object", "created", "owned_by", "type", "display_name", "created_at"],
        "properties": {
          "id": { "type": "string", "examples": ["zencore/space-bunny-free"] },
          "object": { "const": "model" },
          "created": { "type": "integer" },
          "owned_by": { "type": "string" },
          "type": { "const": "model" },
          "display_name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "context_length": { "type": "integer", "description": "Extension: the context window, when the provider gives it" },
          "max_input_tokens": { "type": "integer", "description": "Extension" },
          "max_output_tokens": { "type": "integer", "description": "Extension" }
        }
      },
      "ModelList": {
        "type": "object",
        "required": ["object", "data"],
        "properties": {
          "object": { "const": "list" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Model" } },
          "has_more": { "type": "boolean" },
          "first_id": { "type": "string" },
          "last_id": { "type": "string" }
        }
      },
...
    }
  }
}